refactor: 클래스 이름 변경 및 패키지 구조 수정
변경 내용 :
- Credentials 클래스 이름 AuthInfo 로 변경
- 인증 관련 패키지 생성 및 구분
