+++ src/main/java/cart/controller/CartApiController.java
@@ -1,5 +1,7 @@
 package cart.controller;
 
+import cart.controller.authentication.AuthInfo;
+import cart.controller.authentication.AuthenticationPrincipal;
 import cart.dto.ResponseProductDto;
 import cart.service.CartService;
 import cart.service.MemberService;
@@ -23,22 +25,22 @@ public class CartApiController {
     }
 
     @GetMapping("/carts")
-    public List<ResponseProductDto> readCartItem(@AuthenticationPrincipal Credentials credentials) {
-        final Long memberId = memberService.findIdByEmail(credentials.getEmail());
+    public List<ResponseProductDto> readCartItem(@AuthenticationPrincipal AuthInfo authInfo) {
+        final Long memberId = memberService.findIdByEmail(authInfo.getEmail());
         final List<Long> productIds = cartService.findProductIdsByMemberId(memberId);
         return productService.findByIds(productIds);
     }
 
     @PostMapping("/carts/{productId}")
-    public ResponseEntity<Void> createCartItem(@AuthenticationPrincipal Credentials credentials, @PathVariable final Long productId) {
-        final Long memberId = memberService.findIdByEmail(credentials.getEmail());
+    public ResponseEntity<Void> createCartItem(@AuthenticationPrincipal AuthInfo authInfo, @PathVariable final Long productId) {
+        final Long memberId = memberService.findIdByEmail(authInfo.getEmail());
         cartService.insert(memberId, productId);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/carts/{productId}")
-    public ResponseEntity<Void> deleteCartItem(@AuthenticationPrincipal Credentials credentials, @PathVariable final Long productId) {
-        final Long memberId = memberService.findIdByEmail(credentials.getEmail());
+    public ResponseEntity<Void> deleteCartItem(@AuthenticationPrincipal AuthInfo authInfo, @PathVariable final Long productId) {
+        final Long memberId = memberService.findIdByEmail(authInfo.getEmail());
         cartService.delete(memberId, productId);
         return ResponseEntity.ok().build();
     }
