+++ src/main/java/cart/controller/authentication/AuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.controller.authentication;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
