+++ src/main/java/cart/controller/config/WebMvcConfiguration.java
@@ -1,5 +1,6 @@
-package cart.controller;
+package cart.controller.config;
 
+import cart.controller.authentication.AuthenticationPrincipalArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
